services/scoring-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23

# ML Libraries
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
numpy==1.26.2
pandas==2.1.4

# XAI
shap==0.44.0

# Model Management
joblib==1.3.2
